int改double

+++ src/main/java/ltd/newbee/mall/controller/vo/NewBeeMallIndexConfigGoodsVO.java
@@ -23,7 +23,7 @@ public class NewBeeMallIndexConfigGoodsVO implements Serializable {
 
     private String goodsCoverImg;
 
-    private Integer sellingPrice;
+    private Double sellingPrice;
 
     private String tag;
 
@@ -59,11 +59,11 @@ public class NewBeeMallIndexConfigGoodsVO implements Serializable {
         this.goodsCoverImg = goodsCoverImg;
     }
 
-    public Integer getSellingPrice() {
+    public Double getSellingPrice() {
         return sellingPrice;
     }
 
-    public void setSellingPrice(Integer sellingPrice) {
+    public void setSellingPrice(Double sellingPrice) {
         this.sellingPrice = sellingPrice;
     }
